pyproject.toml
[tool.poetry]
name = "spirit-island"
version = "0.1.0"
description = ""
authors = ["Nathan Jones <nathanj439@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
django-debug-toolbar = "^3.2.4"
Pillow = "^9.0.0"
django-ninja = "^0.16.1"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
"discord.py" = "^1.7.3"
requests-async = "^0.6.2"
structlog = "^21.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
